Move Service to /lib/feature/customattrs
